Add application code
`lib` holds most of the code of the framework itself; the user should never
  have to modify its contents.

`config` holds pre-built files that the user *should* modify -- for instance,
  by defining new partials or view helpers, or specifying routes.

`app` holds user-created (or auto-generated) files specific to the application.
Individual resources in this directory are broken up via a 'pod' architecture.

`index.js` is the entry point, and is responsible for loading the application,
  running any initialization steps, spinning up the server, and managing the
  overall flow of request handling (i.e. Router -> App -> Send Response).