Implement rudimentary version of layouts
Add a 'layouts' directory, and update the view engine to load from there
and interpolate the normal template into it.
Update all existing templates to remove now-redundant <html>, <head>,
and <body> tags.
Update existing partials to make it easier to distinguish between a
layout and the rest of the page's content.