Fix view engine single-quote bug
The view engine was not properly handling the interpolation of
single-quoted strings. It turns out that this was because during the
process of interpolating a template into a layout, the view engine was
calling `.replace`, and `.replace` was mistakenly interpreting `<$'` as
the special replacement pattern `$'`. This issue was not apparent with
double-quoted strings because there is no `$"` replacement pattern. To
fix this issue, update the call to `.replace` to get a function as its
second argument, with that function returning the inner template content
as a string literal.