Implement rough #if block helper
To make this work, treat the content of the block helper as a set of
sub-templates, divided by <$else-if$> and <$else$> tags. Depending on
which condition is true, choose which sub-template to show.

For now, this only works correctly with naked binaries; the next step
would be to add a safe form of evaluation for conditions.