Replace operand string in view helpers with array
Hand view helpers an array of interpreted operands, rather than forcing
each helper to do it for itself.

Note: We are still not performing true evaluation on these operands, and
the parsing system currently does not handle arrays and objects
correctly -- let alone actual expressions. This is an area that will
require extensive further work, since it doesn't seem possible to use
eval() in this context.