Implement basic automated testing tool
Since the components of this framework are getting more complex and
interconnected, it would be a good idea for them to have tests of their
own. However, since the point of this project is to DIY as much as
humanly possible, I thought it would be a neat idea to implement my own
basic testing tool rather than using a stock solution like Mocha or
Jasmine. There's a lot more that can be done here, eventually, but for
now this should be enough to proceed.