Make tests skippable and colorize tester output
Define `xdescribe`, `xcontext`, and `xit` methods on Tester that do
nothing but print the description of the test they would otherwise run.
By creating these methods, any `describe`, `context`, or `it` method
can be skipped simply by adding an 'x' to the front.

Additionally, define methods for printing in color so that it's easier
to visually distinguish between passing, failing, and skipped tests.