Refactor `Parser.extractNextToken`
We need to run through basically the same process of depth-tracking
when dealing with strings, arrays, or objects in order to identify
where a token ends, so spin that out into its own function.

With that done, the rest of `Parser.extractNextToken` becomes simpler,
since now it's only jobs are slicing the token out of the original
string and identifying what type of token it is using RegExp matchers.