Rework `indexOfFirstFreeChar` to match sequences
As it iterates through the larger string, the function will now check
the next N characters (where N is the length of the target sequence)
starting at the current character. If the current character is un-nested
and the next N characters match the sequence, memoize the current index.

Accordingly, rename the function to `indexOfFirstUnnestedMatch`