Define `parseToken` and `parseVariableToken`
Create a function called `parseVariableToken` that reads a token
containing a variable name and attempts to evaluate it, given a specific
context object as a namespace.

Create tests for a method called `parseToken` to figure out how to
handle parse a variety of types of tokens, including 'nullval',
'boolean', 'number', 'string', and 'varName' token types. Then,
implement the method.