Add tests for Parser.identifyTokenType
The rest of the parser code relies heavily on this method, so it doesn't
make sense for it not to have its own tests.

Revisiting this method also led me to realize that there were some cases
I was missing as a result of how variables are being handled. Left a
note to come back to this once everything else is working.