Add evaluation of parentheses in expressions
Change approach to parentheses; rather than treating them as a subset of
functions, treat them as their own distinct operator. This is because
parentheses have much more in common with the other operators than how
functions (and other references, like properties) will need to be
handled. For one thing, both of the latter two need to be checked for
existence before anything can be done. Also the parentheses used in
function calls need to be able to accommodate any number of arguments,
whereas the normal parentheses are much simpler -- just a unary
operator that evaluates its contents and is run with higher
precedence than any of the others.