Update `parseExpressionToken` to handle references
Because `parseToken` already knows how to parse references, there was
really only one significant change that needed to be made to
`parseExpressionToken` to get it to work with references - namely,
making sure that the context object is passed in as an argument every
time `parseToken` is called. There is no need to handle references
independently of the rest of the evaluation process.