Fix `indexOfFirstUnnestedMatch`-related bug
In the event where an unnested match was not found, the helper function
`indexOfFirstUnnestedMatch` was returning `undefined`; however, when
the function is called in `Parser.identifyTokenType`, the result
expected in the event of no match was `null`. `null` is a better
representation of this state, and is easier to work with, so change
`indexOfFirstUnnestedMatch` to return `null` in the event of no match.