Handle expressions containing function calls
Treat function calls as operators; include the function call matcher in
the combined operator matcher, and check for the presence of function
calls when determining of a token is an expression. To parse the
function call, separately capture two tokens: one for the function name
and one for the list of function arguments. The first of these can be
parsed like any other reference; it will return the specified function.
The second, when bracketed by square braces, can be regarded as an array
and parsed as such.

Also, enable parser tests for expressions containing function calls.