Handle string concatenation edge case
This only comes up in the case of parsing expressions, since parsing
expressions is the only time when we might be trying to insert a string
value into the original parent token, so the only part of the code we
need to modify is `parseExpressionToken`. In particular, since this only
comes up with '+', we only need to modify the section dealing with
normal operators.

To determine if string concatenation will be taking place, check to see
if (a) the operator is '+', and (b) at least one of the operands is not
a number. If string concatenation is happening, wrap the resultant
string value in quotes before inserting it into the parent string.