Handle missed edge case in parser
The parser was correctly inserting strings that were the product of
string concatenation back into the parent token. However, this was not
the only way the result of an operation would be a string. One such
missed case is when a function returns a string as a result. To fix
this, update `parseExpressionToken` to check if the result of a function
evaluation is a string (and if it is, wrap it in quote characters).