Rework the handling of inline helpers
Rather than treating inline helpers as non-code entities, instead treat
them as functions that are available in the template's rendering
context. Doing this will erase any distinction between inline helpers
and direct evaluation, since in both cases the entire content of the
tag is an expression to be evaluated.

In addition to changing this behavior, change the syntax slightly from
`<$...$>` to `<$=...$>` in order to more clearly show what is happening
with an inline tag; namely, that it is evaluating an expression. Update
all templates accordingly.

Also, update inline helpers and as necessary to accommodate this change.